move events into subdir
