+++ apps/mobile/src/components/Feed/Events/TokensAddedToCollectionFeedEvent.tsx
@@ -4,37 +4,34 @@ import FastImage from 'react-native-fast-image';
 import { useFragment } from 'react-relay';
 import { graphql } from 'relay-runtime';
 
-import { CollectorsNoteAddedToCollectionFeedEventFragment$key } from '~/generated/CollectorsNoteAddedToCollectionFeedEventFragment.graphql';
+import { TokensAddedToCollectionFeedEventFragment$key } from '~/generated/TokensAddedToCollectionFeedEventFragment.graphql';
 import { removeNullValues } from '~/shared/relay/removeNullValues';
 
-import { Typography } from '../Typography';
-import { EventTokenGrid } from './EventTokenGrid';
-import { FeedEventCarouselCellHeader } from './FeedEventCarouselCellHeader';
-import { FeedListCollectorsNote } from './FeedListCollectorsNote';
+import { Typography } from '../../Typography';
+import { EventTokenGrid } from '../EventTokenGrid';
+import { FeedEventCarouselCellHeader } from '../FeedEventCarouselCellHeader';
 
-type CollectorsNoteAddedToCollectionFeedEventProps = {
+type TokensAddedToCollectionFeedEventProps = {
   isFirstSlide: boolean;
   allowPreserveAspectRatio: boolean;
-  collectionUpdatedFeedEventDataRef: CollectorsNoteAddedToCollectionFeedEventFragment$key;
+  collectionUpdatedFeedEventDataRef: TokensAddedToCollectionFeedEventFragment$key;
 };
 
-export function CollectorsNoteAddedToCollectionFeedEvent({
+export function TokensAddedToCollectionFeedEvent({
   isFirstSlide,
   allowPreserveAspectRatio,
   collectionUpdatedFeedEventDataRef,
-}: CollectorsNoteAddedToCollectionFeedEventProps) {
+}: TokensAddedToCollectionFeedEventProps) {
   const eventData = useFragment(
     graphql`
-      fragment CollectorsNoteAddedToCollectionFeedEventFragment on CollectorsNoteAddedToCollectionFeedEventData {
-        newCollectorsNote
-
+      fragment TokensAddedToCollectionFeedEventFragment on TokensAddedToCollectionFeedEventData {
         collection {
           name
+        }
 
-          tokens(limit: 4) {
-            token {
-              ...EventTokenGridFragment
-            }
+        newTokens {
+          token {
+            ...EventTokenGridFragment
           }
         }
       }
@@ -43,26 +40,20 @@ export function CollectorsNoteAddedToCollectionFeedEvent({
   );
 
   const tokens = useMemo(() => {
-    return removeNullValues(
-      eventData.collection?.tokens?.map((collectionToken) => collectionToken?.token)
-    );
-  }, [eventData.collection?.tokens]);
+    return removeNullValues(eventData.newTokens?.map((collectionToken) => collectionToken?.token));
+  }, [eventData.newTokens]);
 
   return (
     <View className="flex flex-col">
       <FeedEventCarouselCellHeader>
         <Typography className="text-xs" font={{ family: 'ABCDiatype', weight: 'Regular' }}>
-          Added a colloctors note to
+          Added new tokens to
         </Typography>
 
         <Typography className="text-xs" font={{ family: 'ABCDiatype', weight: 'Bold' }}>
           {eventData.collection?.name ?? 'Untitled'}
         </Typography>
       </FeedEventCarouselCellHeader>
-
-      {eventData.newCollectorsNote && (
-        <FeedListCollectorsNote collectorsNote={eventData.newCollectorsNote} />
-      )}
 
       <EventTokenGrid
         imagePriority={isFirstSlide ? FastImage.priority.high : FastImage.priority.normal}
